BACKEND/ai-engine/requirements.txt
# FastAPI for ML API endpoints
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Core ML libraries
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0

# Deep learning and transformers
torch==2.1.0
transformers==4.35.0
sentence-transformers==2.2.2

# Vector database and similarity search
faiss-cpu==1.7.4
chromadb==0.4.18

# LLM APIs
openai==1.3.7
anthropic==0.7.8

# Data processing
requests==2.31.0
aiohttp==3.9.1

# PocketBase client
pocketbase==0.10.1

# Configuration and environment
python-dotenv==1.0.0
pydantic==2.5.0

# Logging
structlog==23.2.0

# Background processing
celery==5.3.4
redis==5.0.1

# Data visualization for insights
matplotlib==3.8.0
seaborn==0.13.0

# Statistical analysis
scipy==1.11.4

# Development and testing
pytest==7.4.3
black==23.11.0
isort==5.12.0